pull main and run precommits

+++ src/clf/xgboost_training.py
@@ -12,17 +12,18 @@ Possible arguments:
 
 """
 import pickle
-import click
 
+import click
 import numpy as np
 import pandas as pd
-from constants import XGB_MIMIC_CLASSIFIED, TRAIN_DATA_DIR
-from imblearn.pipeline import Pipeline
 
 # need to install older version of xgboost for this to work
 import xgboost as xgb
+from imblearn.pipeline import Pipeline
 from sklearn.model_selection import GridSearchCV
-from utils import load_data, preprocessing_pipeline
+
+from src.clf.constants import TRAIN_DATA_DIR, XGB_MIMIC_CLASSIFIED
+from src.clf.utils import load_data, preprocessing_pipeline
 
 
 # Build pipeline
@@ -87,7 +88,7 @@ def grid_search(
     X_train: pd.Series,
     y_train: pd.Series,
     model_pipeline: Pipeline,
-    param_grid: list[dict[str, list[float]]],
+    param_grid: list[dict[str, list[int]]],
 ) -> Pipeline:
     """
     Grid search for best model
